move the friends search into the sidebar
