+++ src/components/main/friends/sidebar/mod.rs
@@ -1,8 +1,24 @@
-use crate::{components::main::friends::request::FriendRequest, Account, LANGUAGE};
 
-use dioxus::prelude::*;
+use crate::{
+    components::{
+        main::friends::{ request::FriendRequest},
+        ui_kit::{button::Button, icon_button::IconButton, icon_input::IconInput},
+    },
+    Account,  LANGUAGE, TOAST_MANAGER,
+};
+
+use arboard::Clipboard;
+use dioxus::{
+    core::UiEvent,
+    events::{FormEvent, MouseData},
+    prelude::*,
+};
+use dioxus_heroicons::{outline::Shape};
+use dioxus_toast::{Position, ToastInfo};
+
 use std::{collections::HashSet, time::Duration};
 use warp::multipass::Friends;
+use warp::crypto::DID;
 
 #[inline_props]
 #[allow(non_snake_case)]
@@ -44,7 +60,7 @@ pub fn Sidebar(cx: Scope, account: Account, add_error: UseState<String>) -> Elem
     cx.render(rsx!(
         div {
             id: "sidebar",
-
+            FindFriends { account: account.clone(), add_error: add_error.clone()},
             div {
                 class: "scroll_wrap",
                 div {
@@ -123,5 +139,133 @@ pub fn Sidebar(cx: Scope, account: Account, add_error: UseState<String>) -> Elem
                 }
             }
         }
+    ))
+}
+
+#[inline_props]
+#[allow(non_snake_case)]
+pub fn FindFriends(cx: Scope, account: Account, add_error: UseState<String>) -> Element {
+    let toast = use_atom_ref(&cx, TOAST_MANAGER);
+    let l = use_atom_ref(&cx, LANGUAGE).read();
+    let remote_friend = use_state(&cx, String::new);
+
+
+    let l2 = l.clone();
+    let codeCopied = { l.code_copied.to_string() };
+    let account2 = account.clone();
+
+    cx.render(rsx!(
+        label {
+            "{l.copy_friend_code}",
+        },
+        div {
+            class: "add",
+            Button {
+                text: l.copy_code.to_string(),
+                icon: Shape::ClipboardCopy,
+                on_pressed: move |e: UiEvent<MouseData>| {
+                    e.cancel_bubble();
+    
+                    let mut clipboard = Clipboard::new().unwrap();
+                    if let Ok(ident) = account2
+                        .read()
+                        .get_own_identity()
+                    {
+                        let single_toast = ToastInfo {
+                            position: Position::TopRight,
+                            ..ToastInfo::simple(&codeCopied)
+                        };
+                        let _id = toast.write().popup(single_toast);
+                        clipboard.set_text(ident.did_key().to_string()).unwrap();
+                    }
+                }
+            }
+        },
+        label {
+            "{l.add_someone}",
+        },
+        div {
+            class: "add",
+            IconInput {
+                placeholder: l.add_placeholder.clone(),
+                icon: Shape::UserAdd,
+                value: remote_friend.to_string(),
+                on_change: move |evt: FormEvent| {
+                    add_error.set(String::new());
+                    remote_friend.set(evt.value.clone());
+                },
+                on_enter: move |_| {
+                    let did = DID::try_from(remote_friend.clone().to_string());
+                    match did {
+                        Ok(d) => {
+                            match account.clone()
+                                .write()
+                                .send_request(&d)
+                            {
+                                Ok(_) => {
+                                    let single_toast = ToastInfo {
+                                        position: Position::TopRight,
+                                        ..ToastInfo::simple(l2.request_sent.clone().as_ref())
+                                    };
+                                    let _id = toast.write().popup(single_toast);
+                                    add_error.set("".into());
+                                    remote_friend.set("".into());
+                                }
+                                Err(e) => {
+                                    remote_friend.set("".into());
+                                    add_error.set(match e {
+                                        warp::error::Error::CannotSendFriendRequest => l2.couldnt_send.to_string(),
+                                        warp::error::Error::FriendRequestExist => l2.already_sent.to_string(),
+                                        warp::error::Error::CannotSendSelfFriendRequest => l2.add_self.clone(),
+                                        _ => l2.something_went_wrong.to_string()
+                                    })
+                                },
+                            };
+                        },
+                        Err(_) => add_error.set(l2.invalid_code.clone()),
+                    }
+                }
+            }
+            IconButton {
+                icon: Shape::Plus,
+                on_pressed: move |e: UiEvent<MouseData>| {
+                    e.cancel_bubble();
+    
+                    let did = DID::try_from(remote_friend.clone().to_string());
+                    match did {
+                        Ok(d) => {
+                            match account.clone()
+                                .write()
+                                .send_request(&d)
+                            {
+                                Ok(_) => {
+                                    let single_toast = ToastInfo {
+                                        position: Position::TopRight,
+                                        ..ToastInfo::simple(&l.request_sent)
+                                    };
+                                    let _id = toast.write().popup(single_toast);
+                                    add_error.set("".into());
+                                    remote_friend.set("".into());
+                                }
+                                Err(e) => {
+                                    remote_friend.set("".into());
+                                    add_error.set(match e {
+                                        warp::error::Error::CannotSendFriendRequest => l.couldnt_send.to_string(),
+                                        warp::error::Error::FriendRequestExist => l.already_sent.to_string(),
+                                        warp::error::Error::CannotSendSelfFriendRequest => l.add_self.to_string(),
+                                        _ => l.something_went_wrong.to_string()
+                                    })
+                                },
+                            };
+                        },
+                        Err(_) => add_error.set(l.invalid_code.to_string()),
+                    }
+                },
+            }
+        },
+        span {
+            class: "error_text",
+            "{add_error}"
+        },
     ))
 }
